pyproject.toml
[tool.poetry]
name = "zero-shot-classification"
version = "0.1.0"
description = "API with zero-shot classification model"
authors = ["igmalta <ignaciomaltagliatti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
python-decouple = "^3.5"
rich = "^10.12.0"
mongoengine = "^0.23.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
transformers = "^4.11.3"
torch = "^1.9.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.1"
isort = "^5.9.3"
pretty-errors = "^1.2.24"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Black formatting
[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
'''

# iSort
[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
skip_gitignore = true
virtual_env = "venv"
